from-submodule now works with recent git submodule implementation
Starting in git 1.7.8, git submodule started storing submodule metadata
in $GIT_DIR/modules/<name>. This broke the git subtree from-submodule
command, which would fail when fetching from the moved repo.

from-submodules now checks to see if the .git file is a file (rather
than a directory), and also moves the actual git metadata into the .git
subdirectory of the moved prefix.

It's also necessary to unset core.worktree when moving the actual git
metadata.

+++ contrib/subtree/git-subtree.sh
@@ -831,7 +831,22 @@ cmd_from-submodule()
 	# Move submodule aside.
 	local tmp_repo="$(mktemp -d /tmp/git-subtree.XXXXX)"
 	rm -r $tmp_repo
-	mv $prefix $tmp_repo
+	if [ -f $prefix/.git ]
+	then
+		# Have to unset core.worktree or this will suffer
+		(cd $prefix && git config --unset core.worktree)
+
+		# Handle new git submodules stored in parent repo
+		local actual_git=$(sed -e 's/^gitdir: //' $prefix/.git)
+		local tmp_git="$(mktemp -d /tmp/git-subtree.XXXXX)"
+		mv $prefix/$actual_git $tmp_git/.git
+		mv $prefix $tmp_repo
+		rm $tmp_repo/.git
+		mv $tmp_git/.git $tmp_repo/.git
+		rm -rf $tmp_git
+	else
+		mv $prefix $tmp_repo
+	fi
 	git rm $prefix
 
 	# Commit changes.
